🐛 Fix : page.tsx 서버컨포넌트 유지를 위한 컴포넌트 분리
